+++ docs/source/changelog.rst
@@ -2445,3 +2445,6 @@ Current C++/SQLite client, Python/SQLAlchemy server
 
 - **New task:** :ref:`Snaith–Hamilton Pleasure Scale (SHAPS) <shaps>`.
   (Database revision 0037.)
+
+- Add optional waist circumference to :ref:`BMI  <bmi>`.
+  (Database revision 0038.)